:construction: create front modules
