+++ front/ui.py
@@ -1,0 +1,17 @@
+from bullet import Bullet
+from bullet import colors
+
+def dropdown(prompt, choices):
+    cli = Bullet(
+        prompt=prompt,
+        choices=choices,
+        indent=0,
+        align=5,
+        margin=2,
+        bullet=">",
+        bullet_color=colors.bright(colors.foreground["yellow"]),
+        word_on_switch=colors.bright(colors.foreground["yellow"]),
+        background_on_switch=colors.background["black"],
+        pad_right=2
+    )
+    return cli.launch()
